requirements.txt
pillow
Flask==2.3.2
frozenlist==1.3.3
huggingface-hub==0.14.1

jsonify==0.5

mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.3
openai==0.27.0

requests==2.31.0
requests-toolbelt==1.0.0
rich==13.3.5
scipy==1.10.1
sqlparse==0.4.4
sympy==1.12
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.29.2
typing_extensions==4.6.0
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.21.0

gunicorn==20.1.0
